fix: rename package to authclaims
